### 2022-09-09 #66 @fbarraquand
[Re] Biodiversity of plankton by species oscillations and chaos
**Original article:** [Huisman, J., & Weissing, F. J. (1999). Biodiversity of plankton by species oscillations and chaos. Nature, 402(6760), 407-410.](https://www.nature.com/articles/46540)

**PDF URL:** [https://github.com/fbarraquand/supersaturated_coexistence/blob/main/article.pdf](https://github.com/fbarraquand/supersaturated_coexistence/blob/main/article.pdf)
**Metadata URL:** [https://github.com/fbarraquand/supersaturated_coexistence/blob/main/Article/metadata.yaml](https://github.com/fbarraquand/supersaturated_coexistence/blob/main/Article/metadata.yaml)
**Code URL:** [https://github.com/fbarraquand/supersaturated_coexistence](https://github.com/fbarraquand/supersaturated_coexistence)

**Scientific domain:** Ecology
**Programming language:** R
**Suggested editor:** Timothée Poisot @tpoisot

Tagging co-authors @DoyenGuillaume @CoraliePicoche
