### 2022-11-10 #66 @tpoisot
**REVIEWER 1** is @swmpkim -- *thank you*!

There is an overview of the review process [here](https://rescience.github.io/edit/). You can post your review as a comment in this thread. If you want to comment on a specific place in the code, you can do it like so:

https://github.com/fbarraquand/supersaturated_coexistence/blob/main/Code/Script_Fig_2.R#L12

But also feel free to voice high-level comments about reproducibility, ask questions, and clarifications. We are expecting the review to be a discussion process aimed at improving the code of the paper!
