### 2022-11-25 #66 @tpoisot
@fbarraquand @DoyenGuillaume @CoraliePicoche - can you suggest reviewers from the [list](https://github.com/rescience/rescience.github.io/blob/sources/04-board.md) or from outside the list?
