### 2022-11-26 #66 @ha0ye
I am familiar with the original Huisman & Weissing paper, but won't have time to do a review until after the next 2 weeks.
