### 2022-11-28 #66 @juergspaak
@fbarraquand I've seen your message. However, I'm not sure how to respond to this. Is this equivalent to a request for a review?
I do know Huisman& Weissing paper, but R is not my primary coding language (but could potentially review this as well).
