### 2022-12-01 #66 @swmpkim
Hi y'all, I'm one of the reviewers and haven't had a chance to look at the paper yet (the new one or the 1999 one; next week, I promise!). But I did run code (R scripts only) and wanted to note a couple of things to update, maybe before the next reviewer comes on board.

I dove into the code without even trying to figure out the ecological context: purely a question of "can I run all this code". And the answer is "almost entirely, yes". The way the files were organized, named, and discussed in the README made it really easy to figure out what to run and when. I appreciated the warning about the code for Fig. 3 in particular taking a long time to run. And best I can tell, the images I generated were the same as the authors'. So with a few updates, yes, the code was reproducible on my machine. Specifics follow.

**Important code updates** - things I had to change in the code to make it run. All notes are for `Script_Experiment`.

- need a call to `library(xtable)` or lines 143 and 144 won't run.
- extra space in line 116, towards the end of the line. There's a `> =` that needs to be `>=`.   Context: `(results_exp1[dim(results_exp1)[1], 2:13])> = 0),rep(T,12)`
- similar extra spaces in lines 135 and 206:  `" = = "` should be `"=="` in the `Probability = c(colSums......)` portion.

**Additional but less critical notes**

- In the README, I believe you're referring to the `save_image()` function here, not `save.image()`: "Finally, saving the plotly plots into .pdf of .png files requires the `save.image()` function."
- It would be helpful with that same line to note that the `save_image()` function requires an installation of python/miniconda, since that particular function isn't available directly in the R {plotly} package. (I didn't understand why I needed the provided installation instructions until running that line in `Script_Fig_1` threw errors.)
- Same function, again - need to make sure the {reticulate} package is recent. ~~The CRAN version that I'd installed had a bug and I had to install the github version in order to get the miniconda installation to work correctly.~~ See [issue 1297](https://github.com/rstudio/reticulate/issues/1297) for details on that issue. ~~Unsure when a new version will be pushed to CRAN.~~  UPDATE: as of 1/8/2023, the version of {reticulate} on CRAN has this bug fixed (see again the linked issue).
- I still never got that function to work but I don't think it's the authors' job to get me through the error I couldn't get past ("NameError: name 'sys' is not defined"). I tried to do some troubleshooting but figured actually saving the images was less important than generating them.
- a couple minor typos in the README, `Code/` .R files section:  `Code/Script_Exp.R` should be `Code/Script_Experiment.R`; `Code/FScript_Functions.R` should be `Code/Script_Functions.R`
- For experiment 1, I got a few of these warning messages; not sure if this matters: `DLSODA-  At T (=R1), too much accuracy requested for precision of machine..  See TOLSF (=R2)
In above message, R1 = 8656.66, R2 = nan`
- Also for experiment 1, I got a few warnings along these lines; again not sure if they are concerning but want to point them out in case.  `1: In lsoda(y, times, func, parms, ...) :
  Excessive precision requested.  scale up `rtol' and `atol' e.g by the factor 10
2: In lsoda(y, times, func, parms, ...) :
  Returning early. Results are accurate, as far as they go
3: In lsoda(y, times, func, parms, ...) :
  an excessive amount of work (> maxsteps ) was done, but integration was not successful - increase maxsteps`

- Didn't get any warnings for experiment 2.

I think that's all for now. I hope it's helpful. Thanks for letting me review this; I'm looking forward to exploring the paper and learning what all this code was for! More next week sometime.
