+++ Code/Script_Fig_appendix.R
@@ -1,0 +1,448 @@
+##########################Used libraries : 
+{
+  library(ggplot2)
+  library(plotly)
+}
+
+##########################Fig_exp_1
+
+stats_exp1 <- read.table("./DataExp/Exp1/stats_exp1.txt")
+stats_exp2 <- read.table("./DataExp/Exp2/stats_exp2.txt")
+
+Nb_Fig <- 3 #Number of additional figures requested
+
+
+set.seed(211232)
+FigList_1 <- c(sample((2:dim(stats_exp1)[1]),7+8*Nb_Fig))[ - (1:7)]# #Random selection of figures that have not yet been displayed
+
+set.seed(81490)
+FigList_2 <- c(sample((2:dim(stats_exp2)[1]),7+8*Nb_Fig))[ - (1:7)]#Random selection of figures that have not yet been displayed
+
+for(k in (1:Nb_Fig)){ #Loop for how much additional figures are requested. Everything inside of it is the same as in Script_Fig_exp_1 & 2, except for the saving part where we change the names of the graphs
+    
+###################Fig_exp_1
+  for(j in (1:8)){
+    dataresults_Fig_exp1 <- read.table(paste("./DataExp/Exp1/dataresults_exp1_",as.character(stats_exp1[FigList_1[k*j], 1]), ".txt", sep = ""))
+    
+    Fig_j1 <- {plot_ly(dataresults_Fig_exp1, x = ~time, y = ~N1, name = "Species 1",  type = "scatter",  mode = "lines", line = list(color = "#FFD700")) %>%
+        add_trace(y = ~N2,  name = "Species 2",  mode = "lines", line = list(color = "#FFFF00")) %>%
+        add_trace(y = ~N3,  name = "Species 3",  mode = "lines", line = list(color = "#7FFF00")) %>%
+        add_trace(y = ~N4,  name = "Species 4",  mode = "lines", line = list(color = "#008000")) %>%
+        add_trace(y = ~N5,  name = "Species 5",  mode = "lines", line = list(color = "#20B2AA")) %>%
+        add_trace(y = ~N6,  name = "Species 6",  mode = "lines", line = list(color = "#0000FF")) %>%
+        layout(title = "The abundances of species 1-6", 
+               xaxis = list(title = "Time (days)", range = c(0, 10000)), 
+               yaxis = list(title = "Abundances"), 
+               shapes = list(list(type = "line", 
+                                x0 = 0.1, x1 = 0.1, xref = "paper", 
+                                y0 = 0,  y1 = max(dataresults_Fig_exp1[,2:7]), 
+                                line = list(color = "black", dash = "dot", width = 1)), 
+                           list(type = "line", 
+                                x0 = 0.3, x1 = 0.3, xref = "paper", 
+                                y0 = 0,  y1 = max(dataresults_Fig_exp1[,2:7]), 
+                                line = list(color = "black", dash = "dot", width = 1)), 
+                           list(type = "line", 
+                                x0 = 0.5, x1 = 0.5, xref = "paper", 
+                                y0 = 0,  y1 = max(dataresults_Fig_exp1[,2:7]), 
+                                line = list(color = "black", dash = "dot", width = 1))
+                           
+               ))}
+    
+    Fig_j2 <- {plot_ly(dataresults_Fig_exp1, x = ~time, y = ~N7, name = "Species 7",  type = "scatter",  mode = "lines", line = list(color = "#8A2BE2")) %>%
+        add_trace(y = ~N8,  name = "Species 8",  mode = "lines", line = list(color = "#000000")) %>%
+        add_trace(y = ~N9,  name = "Species 9",  mode = "lines", line = list(color = "#FF1493")) %>%
[…]
