+++ Code/Script_Experiment.R
@@ -2,11 +2,12 @@
 {
   library(deSolve) 
   library(truncnorm)
+  library(xtable)
 
   source("Script_Functions.R")
 }
 
-##########################experiment 1 : 
+##########################Experiment 1 : 
 {
   
   #size of the test sample : 
@@ -27,9 +28,13 @@
     parameters_exp1<-makeparameters(rep(0, 12), D_exp1, m_exp1, S_exp1, K_exp1, C_exp1)
     
     #randomly choosing the parameter r, but assuring that the firsts r will be the "normal" configuration (only ones)
-    set.seed(1473)
+    set.seed(1151)
     matr_exp1 <- matrix(c(rep(1, 12), rtruncnorm(n_exp1*12, 0.7, 1.3, 1, 0.1)), n_exp1+1, 12, T)
     
+    #saving the values of r  
+    r_exp1 <- cbind((1:(n_exp1+1)), matr_exp1)
+    names(r_exp1) <- c("N°", paste("r", 1:12, sep = ""))
+    write.table(r_exp1, "./DataExp/Exp1/r_exp1.txt")
     
   }
   
@@ -111,41 +116,46 @@
                          results_3_exp1[, 1:13][-dim(results_3_exp1)[1], ], 
                          results_4_exp1[, 1:13])
       
-      #sometimes the simulation can generate error such as negative or NA values.
-      #Before saving the data we check if it's the case or not.
-      if((!is.na(results_exp1[dim(results_exp1)[1], 2]))&(setequal(((results_exp1[dim(results_exp1)[1], 2:13])> = 0),rep(T,12)))){
+      #Sometimes the simulation can generate error such as negative or NA values.
+      #Before saving the data we check if it's the case or not and not processing the data if needed.
+      if((!is.na(results_exp1[dim(results_exp1)[1], 2]))&(setequal(((results_exp1[dim(results_exp1)[1], 2:13]) >= 0),rep(T,12)))){
 
         nametxtj <- paste('./DataExp/Exp1/dataresults_exp1_', as.character(j), '.txt',  sep = "")
         
         write.table(as.data.frame(results_exp1), nametxtj)
     
-        coex_species_1 <- (results_exp1[dim(results_exp1)[1], 2:13]>0.001)
-    
-        stats_exp1 <- rbind(stats_exp1, c(j, matr_exp1[j, ], sum(coex_species_1)))
+        coex_species_1 <- array(0,c(4,12))
+        for (p in (1:4)){
+          coex_species_1[p,] <- (get(paste("results_",p,"_exp1",sep=""))[dim(get(paste("results_",p,"_exp1",sep="")))[1]-1, 2:13]>0.001)
+        }
+        
+        stats_exp1 <- rbind(stats_exp1, c(j,rowSums(coex_species_1)))
       }
     }
     
   }
   
   #saving the statistics 
-  names(stats_exp1) <- c("N°", paste("r", 1:12, sep = ""), "Number of coexisting species")
+  names(stats_exp1) <- c("N°","Number of coexisting species at 1000 days","Number of coexisting species at 3000 days","Number of coexisting species at 5000 days","Number of coexisting species at 10000 days")
   write.table(stats_exp1, "./DataExp/Exp1/stats_exp1.txt")
 
   #displaying it 
-  results_exp1 <- data.frame(Probability = c(colSums(outer(stats_exp1[-1, 14], 0:12, " = = ")), sum(stats_exp1[-1, 14]>5))*(100/dim(stats_exp1[-1, ])[1]))
[…]
