+++ Article/content.tex
@@ -1,9 +1,9 @@
 \section{Introduction}
-In the present article, we replicate the results of Huisman \& Weissing 1999 \textit{Biodiversity of plankton by species oscillations and chaos} \cite{1999:Huisman}, an attempt to resolve the paradox of the plankton \cite{1961:Hutchinson} with a nonlinear ordinary differential equations model based on resource competition theory.\\
+In the present article, we replicate the results of Huisman \& Weissing 1999 \textit{Biodiversity of plankton by species oscillations and chaos} \cite{1999:Huisman}, an attempt to resolve the paradox of the plankton \cite{1961:Hutchinson} with a nonlinear ordinary differential equation model based on resource competition theory.\\
 
 According to many mathematical models, the number of phytoplankton species in a single homogeneous medium cannot exceed the number of separate resources available \cite{1960:Hardin,1973:Phillips,1980:Armstrong}. However it is very common to observe more species than easily identifiable resources in field conditions. This led Hutchinson to formulate the paradox of the plankton \cite{1961:Hutchinson}. Using numerical simulations of their ODE model, Huisman \& Weissing \cite{1999:Huisman} showed that ``supersaturated coexistence'' is possible, where more consumer species than resource items coexist through oscillations or chaos. We attempt here to check this finding. \\
 
-In addition to the replication of the numerical results of Huisman and Weissing \cite{1999:Huisman}, we also present two novel numerical experiments inspired by follow-up articles, \textit{Does ``supersaturated coexistence'' resolve the ``paradox of the plankton'' ?} by Schippers et al. 2008 \cite{2008:Schippers}, and \textit{Towards a solution of the plankton paradox: the importance of physiology and life history} by Huisman et al. 2008 \cite{2008:Huisman}. These articles suggested that supersaturated coexistence might be difficult to obtain outside of the restricted parameter scenarios considered in the original article. Schippers et al. \cite{2008:Schippers} assessed the sensitivity of diversity to parameter perturbations. They did so changing parameters for one species at a time, concluding that supersaturated coexistence was unlikely, while Huisman et al. \cite{2008:Huisman} showed how some trade-offs can make parameter combinations favourable to coexistence emerge (although most of the scenarios they consider confirm the results of Schippers et al. \cite{2008:Schippers}). Here, we further consider mild perturbations of intrinsic population growth rates of all species at once (though not necessarily in the same direction), a hypothesis which we deem coherent with whole-ecosystem perturbations. If supersaturated coexistence is a likely coexistence mechanism, small changes to population growth rates should not massively affect the number of coexisting species. 
[…]
