+++ Article/bibliography.bib
@@ -55,7 +55,7 @@
   journal = {The American Naturalist},
 }
 
-@article{2008:Schippers,
+@article{2001:Schippers,
   doi = {https://doi.org/10.1046/j.1461-0248.2001.00239.x},
   author = {Peter Schippers and Antonie M. Verschoor and Matthijs Vos and Wolf M. Mooij},
   year = {2001},
@@ -67,7 +67,7 @@
   journal = {Ecology Letters},
 }
 
-@article{2008:Huisman,
+@article{2001:Huisman,
   doi = {https://doi.org/10.1046/j.1461-0248.2001.00256.x},
   author = {Jef Huisman and Anna M. Johansson and Eelke O. Folmer and Franz J. Weissing},
   year = {2001},
@@ -81,7 +81,7 @@
 
 @article{2008:Roelke,
   doi = {https://doi.org/10.1086/524955},
-  author={Roelke, Daniel L and Eldridge, Peter M},
+  author={Daniel L Roelke and Peter M Eldridge},
   year = {2008},
   month = February,
   volume={171},
@@ -90,4 +90,18 @@
   title={Mixing of supersaturated assemblages and the precipitous loss of species},
   journal={The American Naturalist}
 }
+
+@article{2015:Edwards,
+	doi = {https://doi.org/10.1890/14-2252.1},
+	author = {Kyle F. Edwards , Christopher A. Klausmeier , Elena Litchman},
+	year = {2015},
+	mounth = {August},
+	volume = {96},
+	number = {8},
+	pages = {2311-2311},
+	title = {Nutrient utilization traits of phytoplankton},
+	journal = {Ecology}
+}
+	
+	
 
