+++ Article/results_exp1.tex
@@ -1,24 +1,0 @@
-% latex table generated in R 4.2.1 by xtable 1.8-4 package
-% Fri Jul 22 14:53:12 2022
-\begin{table}[ht]
-\centering
-\begin{tabular}{rr}
-  \hline
- & x \\ 
-  \hline
-1 species & 13.68 \\ 
-  2 species & 26.56 \\ 
-  3 species & 35.01 \\ 
-  4 species & 11.27 \\ 
-  5 species & 8.85 \\ 
-  6 species & 4.02 \\ 
-  7 species & 0.60 \\ 
-  8 species & 0.00 \\ 
-  9 species & 0.00 \\ 
-  10 species & 0.00 \\ 
-  11 species & 0.00 \\ 
-  12 species & 0.00 \\ 
-  Supersaturated & 4.63 \\ 
-   \hline
-\end{tabular}
-\end{table}
