+++ Code/Script_Fig_exp_2.R
@@ -5,204 +5,230 @@
 }
 
 ##########################Fig_exp_2
-
-stats_exp2 <- read.table("./DataExp/Exp2/stats_exp2.txt")
-
-set.seed(81490)
-FigList <- c(1,sample((2:dim(stats_exp2)[1]),7))#randomly picking among all the simulations
-
-for(j in (1:8)){
-  dataresultsFig_exp2 <- read.table(paste("./DataExp/Exp2/dataresults_exp2_", as.character(stats_exp2[FigList[j], 1]), ".txt", sep = ""))
+{
+  stats_exp2 <- read.table("./DataExp/Exp2/stats_exp2.txt")
   
-  Fig_j1 <- {plot_ly(dataresultsFig_exp2, x = ~time, y = ~N1, name = "Species 1",  type = "scatter",  mode = "lines", line = list(color = "#FFD700")) %>%
-      add_trace(y = ~N2,  name = "Species 2",  mode = "lines", line = list(color = "#FFFF00")) %>%
-      add_trace(y = ~N3,  name = "Species 3",  mode = "lines", line = list(color = "#7FFF00")) %>%
-      add_trace(y = ~N4,  name = "Species 4",  mode = "lines", line = list(color = "#008000")) %>%
-      add_trace(y = ~N5,  name = "Species 5",  mode = "lines", line = list(color = "#20B2AA")) %>%
-      add_trace(y = ~N6,  name = "Species 6",  mode = "lines", line = list(color = "#0000FF")) %>%
-      layout(title = "The abundances of species 1-6", 
-             xaxis = list(title = "Time (days)", range = c(0, 10000)), 
-             yaxis = list(title = "Abundances"))}
+  set.seed(81490)
+  FigList <- c(1,sample((2:dim(stats_exp2)[1]),7))#randomly picking among all the simulations
   
-  Fig_j2 <- {plot_ly(dataresultsFig_exp2, x = ~time, y = ~N7, name = "Species 7",  type = "scatter",  mode = "lines", line = list(color = "#8A2BE2")) %>%
-      add_trace(y = ~N8,  name = "Species 8",  mode = "lines", line = list(color = "#00000")) %>%
-      add_trace(y = ~N9,  name = "Species 9",  mode = "lines", line = list(color = "#FF1493")) %>%
-      add_trace(y = ~N10,  name = "Species 10",  mode = "lines", line = list(color = "#FF0000")) %>%
-      add_trace(y = ~N11,  name = "Species 11",  mode = "lines", line = list(color = "#808080")) %>%
-      add_trace(y = ~N12,  name = "Species 12",  mode = "lines", line = list(color = "#FF8C00")) %>%
-      layout(title = "The abundances of species 7-12", 
-             xaxis = list(title = "Time (days)", range = c(0, 10000)), 
-             yaxis = list(title = "Abundances"))}
+  for(j in (1:8)){
+    dataresults_Fig_exp2 <- read.table(paste("./DataExp/Exp2/dataresults_exp2_", as.character(stats_exp2[FigList[j], 1]), ".txt", sep = ""))
+    
+    Fig_j1 <- {plot_ly(dataresults_Fig_exp2, x = ~time, y = ~N1, name = "Species 1",  type = "scatter",  mode = "lines", line = list(color = "#FFD700")) %>%
+        add_trace(y = ~N2,  name = "Species 2",  mode = "lines", line = list(color = "#FFFF00")) %>%
+        add_trace(y = ~N3,  name = "Species 3",  mode = "lines", line = list(color = "#7FFF00")) %>%
+        add_trace(y = ~N4,  name = "Species 4",  mode = "lines", line = list(color = "#008000")) %>%
[…]
