+++ Code/Script_Fig_exp_1.R
@@ -5,232 +5,259 @@
 }
 
 ##########################Fig_exp_1
-
-stats_exp1 <- read.table("./DataExp/Exp1/stats_exp1.txt")  
-
-set.seed(211232)
-FigList <- c(1,sample((2:dim(stats_exp1)[1]),7))#randomly picking among all the simulations
-
-for(j in (1:8)){
-  dataresults_Fig_exp1 <- read.table(paste("./DataExp/Exp1/dataresults_exp1_",as.character(stats_exp1[FigList[j], 1]), ".txt", sep = ""))
+{
+  stats_exp1 <- read.table("./DataExp/Exp1/stats_exp1.txt")  
   
-  Fig_j1 <- {plot_ly(dataresults_Fig_exp1, x = ~time, y = ~N1, name = "Species 1",  type = "scatter",  mode = "lines", line = list(color = "#FFD700")) %>%
-      add_trace(y = ~N2,  name = "Species 2",  mode = "lines", line = list(color = "#FFFF00")) %>%
-      add_trace(y = ~N3,  name = "Species 3",  mode = "lines", line = list(color = "#7FFF00")) %>%
-      add_trace(y = ~N4,  name = "Species 4",  mode = "lines", line = list(color = "#008000")) %>%
-      add_trace(y = ~N5,  name = "Species 5",  mode = "lines", line = list(color = "#20B2AA")) %>%
-      add_trace(y = ~N6,  name = "Species 6",  mode = "lines", line = list(color = "#0000FF")) %>%
-      layout(title = "The abundances of species 1-6", 
-             xaxis = list(title = "Time (days)", range = c(0, 10000)), 
-             yaxis = list(title = "Abundances"), 
-             shapes = list(list(type = "line", 
-                              x0 = 0.1, x1 = 0.1, xref = "paper", 
-                              y0 = 0,  y1 = max(dataresults_Fig_exp1[,2:7]), 
-                              line = list(color = "black", dash = "dot", width = 1)), 
-                         list(type = "line", 
-                              x0 = 0.3, x1 = 0.3, xref = "paper", 
-                              y0 = 0,  y1 = max(dataresults_Fig_exp1[,2:7]), 
-                              line = list(color = "black", dash = "dot", width = 1)), 
-                         list(type = "line", 
-                              x0 = 0.5, x1 = 0.5, xref = "paper", 
-                              y0 = 0,  y1 = max(dataresults_Fig_exp1[,2:7]), 
-                              line = list(color = "black", dash = "dot", width = 1))
-                         
-             ))}
+  set.seed(211232)
+  FigList <- c(1,sample((2:dim(stats_exp1)[1]),7))#randomly picking among all the simulations
   
-  Fig_j2 <- {plot_ly(dataresults_Fig_exp1, x = ~time, y = ~N7, name = "Species 7",  type = "scatter",  mode = "lines", line = list(color = "#8A2BE2")) %>%
-      add_trace(y = ~N8,  name = "Species 8",  mode = "lines", line = list(color = "#000000")) %>%
-      add_trace(y = ~N9,  name = "Species 9",  mode = "lines", line = list(color = "#FF1493")) %>%
-      add_trace(y = ~N10,  name = "Species 10",  mode = "lines", line = list(color = "#FF0000")) %>%
-      add_trace(y = ~N11,  name = "Species 11",  mode = "lines", line = list(color = "#808080")) %>%
-      add_trace(y = ~N12,  name = "Species 12",  mode = "lines", line = list(color = "#FF8C00")) %>%
-      layout(title = "The abundances of species 7-12", 
-             xaxis = list(title = "Time (days)", range = c(0, 10000)), 
-             yaxis = list(title = "Abundances"), 
[…]
