### 2023-10-04 0f63d0f
Experiment crash and bar plot fixes
