### 2023-11-13 61a2480
New implementation of parameters reading
updates to the codes and to the parameter files
