+++ Code/Matrix_K/K_1ab.txt
@@ -1,3 +1,0 @@
-1 0.75 0.25
-0.25 1 0.75
-0.75 0.25 1
