+++ Code/Matrix_C/C_1c.txt
@@ -1,3 +1,0 @@
-0.04 0.07 0.04 0.10 0.03 0.02
-0.08 0.08 0.10 0.10 0.05 0.17
-0.14 0.1 0.1 0.16 0.06 0.14
