+++ Code/Matrix_K/K_23.txt
@@ -1,5 +1,0 @@
-0.39 0.34 0.30 0.24 0.23
-0.22 0.39 0.34 0.30 0.27 
-0.27 0.22 0.39 0.34 0.30 
-0.30 0.24 0.22 0.39 0.34 
-0.34 0.30 0.22 0.20 0.39
