+++ Code/Matrix_C/C_1ab.txt
@@ -1,3 +1,0 @@
-0.10 0.20 0.15
-0.15 0.10 0.20 
-0.20 0.15 0.10
