+++ Code/Matrix_C/C_4.txt
@@ -1,5 +1,0 @@
-0.04 0.04 0.07 0.04 0.04 0.22 0.1 0.08 0.02 0.17 0.25 0.03
-0.08 0.08 0.08 0.1 0.08 0.14 0.22 0.04 0.18 0.06 0.20 0.04
-0.1 0.1 0.1 0.1 0.14 0.22 0.24 0.12 0.03 0.24 0.17 0.01
-0.05 0.03 0.03 0.03 0.03 0.09 0.07 0.06 0.03 0.03 0.11 0.05
-0.07 0.09 0.07 0.07 0.07 0.05 0.24 0.05 0.08 0.1 0.02 0.04
