+++ Code/Matrix_K/K_1d.txt
@@ -1,3 +1,0 @@
-1 0.75 0.25 0.7 0.2 0.65 0.68 0.38 0.46
-0.25 1 0.75 0.2 1.01 0.55 0.83 1.10 0.85
-0.75 0.25 1 1.10 0.7 0.95 0.6 0.5 0.77
