+++ Code/Matrix_C/C_1d.txt
@@ -1,3 +1,0 @@
-0.1 0.2 0.15 0.05 0.01 0.4 0.3 0.2 0.25
-0.15 0.1 0.2 0.15 0.3 0.35 0.25 0.02 0.35
-0.2 0.15 0.1 0.25 0.05 0.2 0.4 0.15 0.1
