+++ Code/Matrix_K/K_4.txt
@@ -1,5 +1,0 @@
-0.39 0.34 0.30 0.24 0.23 0.41 0.20 0.45 0.14 0.15 0.38 0.28
-0.22 0.39 0.34 0.30 0.27 0.16 0.15 0.05 0.38 0.29 0.37 0.31
-0.27 0.22 0.39 0.34 0.30 0.07 0.11 0.05 0.38 0.41 0.24 0.25
-0.30 0.24 0.22 0.39 0.34 0.28 0.12 0.13 0.27 0.33 0.04 0.41
-0.34 0.30 0.22 0.20 0.39 0.40 0.50 0.26 0.12 0.29 0.09 0.16
