+++ Code/Matrix_K/K_1c.txt
@@ -1,3 +1,0 @@
-1 0.9 0.3 1.04 0.34 0.77
-0.3 1 0.9 0.71 1.02 0.76
-0.9 0.3 1 0.46 0.34 1.07
