+++ Code/Matrix_C/C_23.txt
@@ -1,5 +1,0 @@
-0.04 0.04 0.07 0.04 0.04 
-0.08 0.08 0.08 0.1 0.08 
-0.1 0.1 0.1 0.1 0.14 
-0.05 0.03 0.03 0.03 0.03 
-0.07 0.09 0.07 0.07 0.07 
