### 2023-11-20 6ea3a5e
Removed old article pdf + added mostly style edits to article
