+++ Article/metadata.tex
@@ -12,8 +12,8 @@
 \def \reviewerIINAME{}
 \def \reviewerIIORCID{}
 \def \dateRECEIVED{01 November 2018}
-\def \dateACCEPTED{}
-\def \datePUBLISHED{}
+\def \dateACCEPTED{01 December 2023}
+\def \datePUBLISHED{01 December 2023}
 \def \articleTITLE{[Re] Biodiversity of plankton by species oscillations and chaos}
 \def \articleTYPE{Replication}
 \def \articleDOMAIN{ecology}
@@ -41,4 +41,4 @@
 \author[1]{Guillaume Doyen}
 \author[1,\orcid{https://orcid.org/0000-0002-0867-2130}]{Coralie Picoche}
 \author[1,\orcid{https://orcid.org/0000-0002-4759-0269}]{Frédéric Barraquand}
-\affil[1]{Institut de Mathématiques de Bordeaux, CNRS UMR5251, Bordeaux, France}
+\affil[1]{Institut de Mathématiques de Bordeaux (UMR 5251), University of Bordeaux, CNRS, Bordeaux INP, Talence, France}
