+++ Article/metadata.yaml
@@ -32,8 +32,8 @@ authors:
 # contributions" as name (with no address).
 affiliations:
   - code:    1
-    name:    Institut de Mathématiques de Bordeaux, CNRS UMR5251
-    address: Bordeaux, France
+    name:    Institut de Mathématiques de Bordeaux (UMR 5251), University of Bordeaux, CNRS, Bordeaux INP
+    address: Talence, France
 
 # List of keywords (adding the programming language might be a good idea)
 keywords: ecological modelling, plankton, oscillations, coexistence, R
