### 2023-12-04 57ca42b
Added revision files, small tweaks to language and figure width
