+++ Article/metadata.tex
@@ -11,9 +11,9 @@
 \def \reviewerIORCID{}
 \def \reviewerIINAME{}
 \def \reviewerIIORCID{}
-\def \dateRECEIVED{01 November 2018}
-\def \dateACCEPTED{01 December 2023}
-\def \datePUBLISHED{01 December 2023}
+\def \dateRECEIVED{XXX}
+\def \dateACCEPTED{XXX}
+\def \datePUBLISHED{XXX}
 \def \articleTITLE{[Re] Biodiversity of plankton by species oscillations and chaos}
 \def \articleTYPE{Replication}
 \def \articleDOMAIN{ecology}
