+++ Article/metadata.yaml
@@ -99,9 +99,9 @@ contributors:
 
 # This information will be provided by the editor
 dates:
-  - received:  November 1, 2018
-  - accepted:
-  - published: 
+  - received:  XXX
+  - accepted:  XXX
+  - published: XXX
 
 # This information will be provided by the editor
 article:
