### 2023-12-04 #66 @fbarraquand
Hello @tpoisot

We have now revised the manuscript ([pdf](https://github.com/fbarraquand/supersaturated_coexistence/blob/main/article.pdf), revised files have been pushed to the main branch). The response letter is attached to this comment in pdf format (too long for github comments). We are grateful to the two reviewers for constructive criticism.

[doyen_et_al_response_letter_rescience.pdf](https://github.com/ReScience/submissions/files/13550201/doyen_et_al_response_letter_rescience.pdf)
