### 2023-12-04 #66 @rougier
@tpoisot if you need help for the publication, just ask me.
