### 2023-12-05 #66 @rougier
I can take car of the actual publication. @fbarraquand where are your latex sources?
