### 2023-12-06 #66 @DoyenGuillaume
Hi @rougier

If by latex sources you are referring to the `.tex` files used to generate `article.pdf`, you can find them in  `Article` folder of the repository.

BR
