### 2023-12-06 #66 @rougier
@tpoisot did you fill the metadata.yaml file by any chance? If not I'll do it.
