### 2023-12-06 #66 @tpoisot
I did it on the version I put in the articles repo for publication, but I'm not sure that I pushed the modifications. I can do it again later today.
