### 2023-12-06 #66 @rougier
Great, tell me when it's done (or just post the metadata file here)
