### 2023-12-08 #66 @DoyenGuillaume
Hello !

Since everyone is sharing his orcid, I made one for myself as well. Here it is : https://orcid.org/0009-0008-6401-9022
Thanks again to the reviewers, editors and publishers for your feedback on our work.

Br
