### 2023-12-13 #66 @rougier
@tpoisot Thanks
@juergspaak No need to review again! (and many thanks for your review)
@DoyenGuillaume Great. I'll complete the metadata and published the paper first on sandbox zenodo and then zenodo. Stayt tuned!
