### 2023-12-13 #66 @rougier
Here is the sanbox version: https://sandbox.zenodo.org/record/8397

@DoyenGuillaume Can you check if everything looks ok? Waiting for your go to publish the final version.
