### 2023-12-13 #66 @rougier
It's online: https://zenodo.org/record/10371655. It'll appear soon on the journal website.
